Fix certificate errors when accessing koji-hub
During a build, koji-hub is accessed from outside the
default network created by docker-compose. For this reason
the hostname "koji-hub" does not resolve as expected. The
alternative is to access koji-hub by its IP address, which
has worked in the past. However, due to newer koji versions
certificate verification is stronger, requiring the IP address
to also be list in koji-hub's identity certificate.

For simplicity the special IP address 172.17.0.1 is used for
certificate generation. This IP has shown to be consistent
across different hosts (container registry is already accessed
this way).